Align with RFC 6763
Bring mdns_lite into conformance with
RFC 6763 - DNS Service-based discovery.
SRV and PTR now respond correctly.

PTR response is dependent on domain
argument. If all services are asked for,
respond with multiple PTR records.
If request is for a paticular service,
respond with a PTR, TXT, A records
for that service.
If request is a reverse domain lookup,
respond with a PTR record.

An SRV request for a host's service
responds with  SRV and  A records.