Reduce churn on responders when IPs change
The previous logic stopped all old responders and then started
everything in the new list even if an responder was in both. This
changes the logic so responders are on interfaces that didn't change
don't get restarted.

This does two things: 1. it reduces SASL logs about GenServer
stop/starts that look problematic, but aren't, and 2. it allows
responders to keep state. The latter will be useful when record caching
is added.