Use the new OTP DNS support for mDNS messages
This fixes issues with handling the unicast bit when using OTP's
built-in DNS record decoder. Before Erlang/OTP 24.1.2, the DNS decoder
decoded TXT and SRV records even if it didn't recognize the record's
class. Since the mDNS's unicast response bit uses the high bit of the
class field, this was convenient. OTP 24.1.2 fixed some issues and ended
up getting pickier on what it encoded and decoded. On the bright side,
the mDNS unicast response bit is supported now.

This commit updates MdnsLite to support the new :inet_dns updates.
Rather than trying to deal with pre-OTP 24.1.2 behavior, the OTP 24.1.2
behavior and the new support, this commit vendors the encode and decode
code.

All DNS message encoding and decoding is routed through `MdnsLite.DNS`
now. Thorough DNS testing of the decoder and encoder is in OTP. This
contains spot checks for mDNS.

Fixes #59.