Fix DNS bridge failure due to unpreserved rd field
OTP 25's DNS resolver checks that the rd field is preserved in DNS
responses. It turns out that that DNS bridge was not preserving it and
this caused its responses to be discarded.

This fixes the issue. It also preserves opcode and sets rcode. Both of
these default to the right value, but if OTP ever changes the default
mdns_lite shouldn't break.