IPv6 Query Support
This adds support for querying mDNS over IPv6. When IPv6 is enabled, a
responder is started for each of an interface's IPv6 addresses in
addition to any IPv4 addresses.

Responders bound to IPv6 addresses currently do not answer any incoming
queries, so IPv6-only interfaces will still not be able to advertise any
services.